Added checks to keep all network data nodes in sync
Added checks to keep all network data nodes in sync, this way after every round they will check if there is a majority. If there is not they will sync from the main network data node, If there is they will sync from a random network data node if they are not in the majority.